% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpuscomp-package.R
\docType{package}
\name{corpuscomp-package}
\alias{corpuscomp}
\alias{corpuscomp-package}
\title{corpuscomp: STATS 290 Project: Analyzing Corpus Data at the Character Level}
\description{
This package takes in corpus-level data: in particular, it develops analytical frameworks to help users analyze trends in character representation for the constituent documents within a corpus. Given a preprocessed set of documents (e.g., parses generated by the NLP pipeline, "BookNLP"), it helps users wrangle, visualize and document how characters are characterized within their corpus of literary texts.
}
\author{
\strong{Maintainer}: Kate Li \email{kateli@stanford.edu} [copyright holder]

Other contributors:
\itemize{
  \item Sang Truong [contributor, medium]
}

}
\keyword{internal}
