% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpus.R
\name{getbook}
\alias{getbook}
\title{This function essentially applies the create_character_data function to all files
within your corpus, extracting only the .book files}
\usage{
getbook(filepath, outputpath = NULL)
}
\arguments{
\item{filepath}{Filepath containing .book files (scripts parsed by BookNLP on corpus)}

\item{outputpath}{Filepath under which output will be stored}
}
\value{
Returns a list of dataframes, where each dataframe corresponds to atttributes of all characters within a text
}
\description{
This function essentially applies the create_character_data function to all files
within your corpus, extracting only the .book files
}
