% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_funcs.R
\name{grab_collocates}
\alias{grab_collocates}
\title{Analyze Collocates in a Corpus}
\usage{
grab_collocates(filepath, terms, horizon)
}
\arguments{
\item{filepath}{Character. The path to the directory containing \code{.tokens} files.}

\item{terms}{Character vector. The terms for which collocates are identified.}

\item{horizon}{Integer. The number of words before and after the target term to include in the window (default: 5).}
}
\value{
A data frame containing collocates for each term, along with their counts and significance p-values.
}
\description{
This function identifies collocates of specified terms in a corpus by searching
for words that occur within a specified window (\code{horizon}) around the target terms.
}
