% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdb.R
\name{grab_racialmods}
\alias{grab_racialmods}
\title{Grabbing Racial Modifiers -----}
\usage{
grab_racialmods(example_text, db = "rsdb")
}
\arguments{
\item{example_text}{A dataframe containing character, such as one entry in the list output passed from "getbooks."}

\item{db}{A referential database containing a glossary of racializing terms. This currently links to the built in dataframe, "rsdb."}
}
\value{
A dataframe containing the character, the racializing term used to describe them (if any),
and the number of times this modifier was used.
}
\description{
This function grabs the modifiers used to describe the characters in a single text
(e.g., one dataframe from the list output of "getbooks").
It widens the dataframe for further wrangling, so that users can see whether/how many
racializing/ethnographic terms (from the built-in dataframe "rsdb") are used
as a modifier to describe any of the characters in a given parsed text.
Note that this function takes in a single text at a time:
it is not built to be passed onto an entire corpus.
}
