% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{grab_sentiment}
\alias{grab_sentiment}
\title{Sentiment Analysis of Characters -----}
\usage{
grab_sentiment(parsedbooks, colname, method = "syuzhet")
}
\arguments{
\item{parsedbooks}{A output (type list) generated from running getbook() on a corpus.}

\item{colname}{The variable (column) in parsedbooks upon which sentiment analysis is conducted.
Takes on one of four values ("Actions Done", "Actions Done On", "Possessions", or "Modifiers").}

\item{method}{The method of sentiment analysis to be conducted.}
}
\value{
A list of modified dataframes containing character data for each text.
}
\description{
This function takes in a list of character dataframes (e.g., a list output produced by running
"getbooks" on BookNLP-parses), and outputs a modified list of dataframes, in which additional
variables describe the average sentiment of a set of words associated with a given character
(e.g., actions done on/by, modifiers of the character, etc.), as well as the standard deviation
of the sentiments associated with that set of words (e.g., the spread of sentiment used
to characterize a character).
Note that a "positive" sentiment score indicates that a character is described positively,
and a "negative" sentiment score indicates that a character is described negatively.
}
