% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment.R
\name{plot_sentiment}
\alias{plot_sentiment}
\title{Plotting Character Sentiment ---------}
\usage{
plot_sentiment(sentiment_list, measure, charalevel)
}
\arguments{
\item{sentiment_list}{The output of a grab_sentiment run. Takes on the form of a list of dataframes, where each df covers a document's characters and the sentiments they are associated with}

\item{measure}{Either "mean" or "sd." Represents which measure of sentiment is plotted in the resulting visualization (the spread of, or average, sentiment of the character).}

\item{charalevel}{The number of characters whose descriptions to return. (1 corresponds to returning only the major character; 10 corresponds to returning the 10th most major characters.)}
}
\value{
A ggplot. Note that the x-axis can (and likely should) be constrained to fit a smaller number of characters than there are total in the corpus.
}
\description{
This function plots the sentiments associated with characters for each document within a corpus.
In particular, it looks at either the average or spread of sentiment for each character,
in relation to that character's majorness or minorness. It effectively allows for users to
visualize general changes in how characters are described as they become more or less major.
}
