% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corpus.R
\name{proc}
\alias{proc}
\title{For usage in the "getbook" function.
This, very simply, reads in a JSON file.}
\usage{
proc(filename)
}
\description{
For usage in the "getbook" function.
This, very simply, reads in a JSON file.
}
