% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsdb.R
\docType{data}
\name{rsdb}
\alias{rsdb}
\title{Racial Slur Database (rsdb)}
\format{
A data frame with 3 variables:
\describe{
\item{Slur}{variable of type character listing racial slurs.}
\item{Represents}{variable of type character describing what the slur represents.}
\item{Reason & Origins}{variable of type character describing the reasons and origins to the slurs}
}
}
\usage{
data(rsdb)
}
\description{
A dataset containing examples of racial slurs and their representations.
}
\keyword{datasets}
